Allow allocating from an alternate device
Allow specifying an alternate device as the
surface allocator.  This is useful both for
testing cross-device allocation sharing and for
testing the NVIDIA binary driver which currently
allocates from the NVIDIA kernel driver API and
device files rather than the DRM API.

This option only affects the behavior of the
generic unix allocator path.  It is not supported
on the GBM path.

Signed-off-by: James Jones <jajones@nvidia.com>
Reviewed-by: Miguel A Vico Moya <mvicomoya@nvidia.com>